requirements.txt
bcrypt==4.2.0
cffi==1.17.0
cryptography==43.0.0
paramiko==3.4.1
prometheus_client==0.20.0
pyasn1==0.6.0
pycparser==2.22
PyNaCl==1.5.0
pysmb==1.2.9.1
PyYAML==6.0.2

